+++ src/components/manage/Tiles/HeroImageLeft/Edit.jsx
