+++ src/components/manage/Tiles/HeroImageLeft/Edit.test.jsx
