+++ src/components/manage/Tiles/HeroImageLeft/View.jsx
