+++ src/components/manage/Tiles/HeroImageLeft/View.test.jsx
